requirements.txt
pypng
pyqrcode
reportlab
requests
Pillow
PyPDF2
natsort
scikit-learn
numpy